refactor: Counsel - DTO Service Module로 이전
Service, Controller
- DTO package 변경
